fix: only allow non-shared uint8array
Typescript recently changed the default of `Uint8Array` to basically be
`Uint8Array<SharedArrayBuffer | ArrayBuffer>`. This broke all sorts of
things, as expected.

It now means consuming code of nanotar needs to cast the result to
`Uint8Array<ArrayBuffer>` to be able to pass it into anything useful,
basically.

We can fix this by changing all our usages of `Uint8Array` to be
explicitly of type `ArrayBuffer`, since we shouldn't ever be producing
or consuming shared array buffers for this stuff.